refactor: make cryptographic toolbox standalone
